Refactored HUD, Character
There is now a new type of Actor, the Character. (This was planned, as
you can see from the naming of player/var/actor/character)
Character are actors that players can connect to, and have player
centered features. As such, the HUD now belongs under /character.

Also, removing a /tile from your HUD now stops your current action if
that action was using that /tile.